Edits to the README (#3)
* Remove instructions about asdf and Homebrew

Candidates may already have some Ruby version manager of their
liking. They should not be required to use asdf or brew for a
test like this. Also, they may be running on Linux too.

I think that being able to install things like these should be a
given in any senior candidate. If they need guidance or struggle
with this, that's already a signal in my view.

* Remove version requirement for Postgres

Candidates may have a different version of Postgres already installed
and they should not go to the trouble of running one specific version
just for this test.

The application is simple, it does not exploit recent Postgres features.

* Fix spelling of Node.js

Details are important, and correct spelling shows care for them, and
also respect for the spelling chosen by project authors.

* Document bin/setup and seed the db there

* Run `heroku` with `bundle exec`

Otherwise you may get

  Gem::LoadError: You have already activated rake 13.0.3, but your
  Gemfile requires rake 13.0.1. Prepending `bundle exec` to your
  command may solve this.

* ./bin/rspec -> bin/rspec

For consistency with other commands.

* Update Copyright

* Update bin/setup

Co-authored-by: Miha Plohl <kitschmaster@gmail.com>

Co-authored-by: Miha Plohl <kitschmaster@gmail.com>