[Migration] Fill on_shelf data from view
Data migration to fill on_shelf data from
`product_on_shelf_quantities` view

During this migration, we take a SHARE LOCK
on the tables that are used in the view so that
the view data doesn't become stale during the
migration. Only WRITES are locked, READS can be done.

The downtime due to this will depend upon the number of
products in the system which shouldn't be extremely high
Hence, the effect should be negligible.