[Migration] Add the support to have roles for employees
Currently we have requirement for 2 roles - warehouse employees
and customer_service employees. So it's good to have an enum
to keep different roles.
Having an enum also makes it easy to have any new roles in the
future.

There is a standard way of defining enums in rails on the application
level with integer fields in Database.

However, having an enum
on DB level (and then on app level - see next commit) depicts a little
bit more correctness and readabilityin (inside DB) as well.
Also, the other enum `inventory_statuses` has been defined in DB
so I want to be consistent with that.

We also update all existing employees role to be `warehouse` as this
is one of the jobs to be done.