[Feature] Added the ability to move returned inventory to shelf
What?
=====
Once an order is returned, the inventory is not automatically
moved to shelf.
This commit gives the capability to a warehouse employee
to return inventory on shelf on a per-product basis from the portal

Changes?
========
A service, route, controller and UI has been added to move inventory on shelf
Service specs and feature specs to test the complete flow
have also been added